Split client into tywo versions: Tcp and Tls
Create Connect trait to abstract bind and connect method.
Define TcpConnector and TlsConnector implementing trait Connect, each representing a different kind of connection: plain Tcp stream or TLS stream wrapping a Tcp one.
Use generics in Server definition.
Modify CLI to handle the case in which TLS is used or not.
Client is now a generic struct whose internal element implement a ClientConnect trait. This trait is necessary to make the client connect either with or without TLS, depending on the fact that the CLI invokes the method with_tls().
SenderClient has been renamed in Client, the same for the SenderClientError enum.

Signed-off-by: rgallor <riccardo.gallo@secomind.com>