Start working at the message protocol.
we are using a protocol to handle the sending and receiving of messages.
The protocol is defined in the `protocol.rs` file, which includes two enums
that describe the messages a host can send to a device and the messages that a
host can expect to receive from a device. It also includes a Tower Service and
Layer to implement it into the services' stack. The `handle_connection` method
of the `HandleConnection` struct has been updated to accommodate these new data structures.
Please note that the protocol is currently under development and is not yet in its final form.

Signed-off-by: rgallor <riccardo.gallo@secomind.com>