Implement Protocol layer.
The host and the device now communicate through a layer based on two structs,
HostMsg and DeviceMsg, describing the type of messages the device and the
host respectively expect during a communication.
Serde is used to serialize and deserialize from/to the Message type.
The protocol simplifies the code, allowing to remove the channels responsible for the
error and command outputs handling, making it possible to avoid using multiple tokio tasks.

Signed-off-by: rgallor <riccardo.gallo@secomind.com>